server/backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-cors-headers==3.5.0
django-mysql-rds==0.3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
mysqlclient==2.0.1
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
